+++ config/key_bindings.py
@@ -1,17 +1,25 @@
+"""
+可以在这里自定义按键
+
+但是在程序的运行过程中，它们不应该被改变，所以使用大写字母作为名称
+"""
 from pygame.locals import *
 
-up = K_UP
-down = K_DOWN
-left = K_LEFT
-right = K_RIGHT
-confirm = K_z
-cancel = K_x
-cheat = K_c
-restart = K_r
-save = K_s
-load = K_l
-quit = K_q
-show_help = K_h
+UP = K_UP
+DOWN = K_DOWN
+LEFT = K_LEFT
+RIGHT = K_RIGHT
+CONFIRM = K_z
+CANCEL = K_x
+CHEAT = K_c
+RESTART_GAME = K_r
+SAVE_GAME = K_s
+LOAD_GAME = K_l
+EXIT_GAME = K_q
+DISPLAY_HELP = K_h
+RETURN_TO_TITLE = K_ESCAPE
+TOGGLE_BGM = K_b
+TOGGLE_SOUND_EFFECT = K_e
 
 # game specific: card prisoner
-draw_card = K_d
+DRAW_CARD = K_d